+++ src/components/StepWizard/index.js
@@ -1,0 +1,1 @@
+export { default as StepWizard } from './StepWizard'